Fix the file structure for healthz handlers in Hono
