+++ front-api/routes/healthz/startup.ts
@@ -1,54 +1,9 @@
 import { getRedisHybridManager } from "@app/lib/api/redis-hybrid-manager";
-import { COMMIT_HASH } from "@app/lib/commit-hash";
 import { frontSequelize } from "@app/lib/resources/storage";
-import { isInShutdown } from "@app/lib/shutdown_signal";
 import { getStatsDClient } from "@app/lib/utils/statsd";
 import logger from "@app/logger/logger";
 import { normalizeError } from "@app/types/shared/utils/error_utils";
 import { Hono } from "hono";
-
-export const healthzApp = new Hono();
-
-healthzApp.get("/", (ctx) => {
-  const startMs = performance.now();
-  const response = ctx.text("ok", 200);
-  const elapsedMs = performance.now() - startMs;
-
-  getStatsDClient().distribution("requests.health.check", elapsedMs);
-
-  return response;
-});
-
-/**
- * Readiness probe endpoint.
- *
- * This endpoint is checked continuously by Kubernetes to determine if the pod should receive
- * traffic. It's kept simple and doesn't check dependencies to avoid marking all pods unready if
- * Redis/DB have transient issues.
- *
- * During pod shutdown, this probe fails immediately to signal the load balancer to stop
- * sending new connections and begin connection draining.
- *
- * The startup probe (/api/healthz/startup) handles dependency checking at pod startup.
- */
-healthzApp.get("/ready", (ctx) => {
-  const startMs = performance.now();
-
-  if (isInShutdown()) {
-    const durationMs = performance.now() - startMs;
-    getStatsDClient().distribution("healthz.ready.duration_ms", durationMs);
-    getStatsDClient().increment("healthz.ready.shutdown");
-
-    return ctx.json({ status: "shutting_down" }, 503);
-  }
-
-  const response = ctx.json({ status: "ready", commitHash: COMMIT_HASH }, 200);
-
-  const durationMs = performance.now() - startMs;
-  getStatsDClient().distribution("healthz.ready.duration_ms", durationMs);
-
-  return response;
-});
 
 const DEPENDENCY_CHECK_TIMEOUT_MS = 2000;
 
@@ -94,7 +49,9 @@ async function checkDependency(
  * accepting traffic. It's called during pod startup only.
  *
  */
-healthzApp.get("/startup", async (ctx) => {
+const app = new Hono();
+
+app.get("/", async (ctx) => {
   const startMs = performance.now();
 
   const results = await Promise.all([
@@ -138,3 +95,5 @@ healthzApp.get("/startup", async (ctx) => {
     503
   );
 });
+
+export default app;
